==> codi/3_RESULTATS_METPLUS.Rmd <==
---
title: 'eControl Met +: Study about the results of the addition of a sulfonylurea, DPP4 inhibitors or SGLT2 inhibitors as a second antidiabetic drug in patients with diabetes mellitus type 2 in treatment with metformin and insufficient glycemic control. Analysis: *`r params$analisis`*'
author: "Jordi Real"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
  pdf_document: default
  word_document: default
params: 
  analisis: "Intention to treat (ITT)"
  mostra: "total" # "total" / "test"
  analisis_OT: TRUE
  
---


```{r setup, echo=F, include=F}

knitr::opts_chunk$set(echo = FALSE, message=FALSE, warning=FALSE, include=F,size="huge")

# rm(list=ls())
# memory.size(max=160000)

# Parametres
# output_Rdata<-here::here("resultats","Output_metplus.RData")
conductor_variables<-here::here("variables_metplus.xls")

# Càrrega de libreries i funcions  --------------
library(compareGroups)
library(broom)
library(kableExtra)
library(ggplot2)


#   SOURCE
# CArrego funcions -------------------
link_source<-paste0("https://github.com/jrealgatius/Stat_codis/blob/master/funcions_propies.R","?raw=T")
devtools::source_url(link_source)

source(here::here("codi","global_metplus.R"))

if (params$mostra=="test") dir_resultats<-"resultats/test" else dir_resultats<-"resultats"


# Canviar nom d'output 
if (params$analisis_OT) output<-"Output_metplusOT.Rdata" else output<-"Output_metplusITT.Rdata"


# directori_figures  
if (params$analisis_OT) dir_fig<-"resultats/figures_OT" else dir_fig<-"resultats/figures_ITT"

if (params$mostra=="test") dir_fig<-dir_resultats


```



```{r filtres, include=F}

# Carrega Rdata
load(here::here(dir_resultats,output))

dt_codindex<-readRDS(here::here("dades/preparades","dt_codindex.rds")) %>% 
  semi_join(mutate(dades_completes,idp=as.character(idp)),by="idp")

dades<-etiquetar(dades,taulavariables = conductor_variables)



```



```{r generacio_taules, echo=F, include=F}


# 10.1 Taules descriptives exploratories  ----------------
taules<-llistadetaules.compare(tablero=c("table1","table2","table3","table4"),y="grup",variables = conductor_variables,dt=dades_completes)

formula<-formula_compare(x="table1",y="grup",taulavariables = conductor_variables)
taula1.post<-descrTable(formula,data=dades_completes,show.p.overall = F)

# # 10.2.Taula events -----------
# formula<-formula_compare(x="table5",y="grup",taulavariables = conductor_variables,dt=dades_completes)
# taula_events<-descrTable(formula,data=dades, show.ratio = T)


```

# Study Objectives and Endpoints:

## The primary objective:

To compare the proportion of patients achieving the reduction in HbA1c values of at least 0.5%, a weight reduction of at least 3%, after the addition of a SU, an DPP-4i or an SGLT-2i to the treatment with metformin in patients with T2DM and insufficient glycemic control in the medium-long term, up to a maximum of 24 months of follow-up. 

***

## Secondary objectives: 

- To estimate the mean reduction of HbA1c and body weight separately after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control (up to 24 months).
- To assess the average reduction in systolic blood pressure after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control.
- To assess the evolution of lipid profile (mean total cholesterol, HDL, LDL, and triglycerides) after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient control glycemic. 
- To assess  the adherence to treatment after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control.
- To assess  the percentage of suspensions/dropouts of treatment at 6, 12 and 24 months (persistence of treatment).
- To describe the relevant adverse reactions produced after the addition of an SU, an DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control.
- To describe the comorbidity profile and baseline clinical characteristics of patients with DM2 and insufficient control with metformin in monotherapy after addition of a treatment with SU, DPP-4i or SGLT-2i.
- To describe the clinical characteristics of patients who have the best efficacy results (patients with a reduction of HbA1c of more than 0.5% and a weight reduction of at least 3% before the end of follow-up).

***

## Estado:

### Últimas actualizaciones 

> Novembre 2020

&check; Frecuencia de fármaco índice por grupo (Resultado 8)  <br/>

### Realizado

&check; Añadir HbA1c,(mmol/mol),reduction, mean (SD)  <br/>

> Agost 2020

&check; Generar 6 variables nuevas IFCC HbA1c (nmmol/mol) = (10.93*HB1AC)-23.50  <br/>
&check; Arreglar punts de tall d'outcomes reduccions categoriques >= 3% Pes i >=0.5 glicada  <br/>
&check; Generar nou outcome binari <7% HBA1C seguiment 3-24 mesos <br/>
&check; Noves mesures d'associació nou outcome binari <br/>
&check; Canvi de categoria de referencia ISGLT2 en adherencia <br/>
&check; Afegida descriptiva de temps de dispensació / prescripció  <br/>
&check; Actualitzar resultats OT i ITT  <br/>

> Juny 2020

&check; Diferencias estandarizadas pre-post <br/>
&check; Figuras pre post matching  <br/>
&check; Effect sizes por cada outcome 2 a 2 <br/>
&check; Forest plot <br/>
&check; Recalcular IC98% <br/>
&check; Analisis OT <br/>

>

&check; Generar nuevo agregador CVD  <br/>
&check; Eliminado codigos T14 de agregador (musculoskeletal system adverse reactions)  <br/>
&check; Diferencia de PAS entre grupos: SU, ISGLT2  <br/>
&check; Cálculo de p's de diferencias en las reacciones adversas  <br/>
&check; Revisado código.  <br/>
&check; Incluir tabaco/PAS/PAD en tablas pre matching.  <br/>
&check; Añadir comparativa tomando categoria de referencia grupo ISGLT2  <br/>
&check; Marginales totales de algunas tablas   <br/>
&check; Añadir medianas + IQR de cuantitativas  <br/>
&check; Cálculo de p-valores faltantes <br/>
&check; Añadido número de envases por grupo durante el seguimiento <br/>
&check; Añadido tiempo medio de reducción de valor principal durante el seguimiento <br/>
&check; Reselección de muestras (nuevo matching PS) añadiendo más variables no es establecidas en el protocolo inicial (DM duration + CKDEPI) <br/>
&check; Validación de nuevo matching PS  <br/>
&check; Generación de nuevos resultados <br/>
&check; Generació de muestras Imputación múltiple de missings  <br/>
&check; Restimación de parámetros mediante Imputación múltiple de missings  <br/>
&check; Reajuste de estimaciones <br/>
&check; Análisis tomando como referencia resto de grupos <br/>
&check; Análisis tomando como referencia cada uno de los grupos <br/>
&check; Añadir N marginal en tabla para evaluar missings <br/>
&check; Reestructuración de tablas según protocolo <br/>
&check; Verificado código: no inclusiones de metformina en Sulfunurea <br/>
&check; Generado gráfico de verificación de seguimiento de fármacos <br/>
&check; Revisado y arreglado código de análisis On Treatment (No cuadrava con ITT)  <br/>
&check; Ampliado gap de discontinuidad de 61 a 92 dias <br/>
&check; Programación funcional (LPA-I) de estudio segun diseño <br/>
&check; Lectura de ficheros historicos <br/>
&check; Generación de fecha índice  <br/>
&check; Generación de fecha de fin de tratamiento y/o cambios <br/>
&check; Agregación de variables en fecha índice <br/>
&check; Aplicado criterios de exclusión y generación de Flow-charts <br/>
&check; Generación de muestras y emparejamiento <br/>
&check; Validación de matching <br/>
&check; Cálculos y recodificaciones de nuevas variables <br />
&check; Analisis según objetivos  <br />

### Pendiente

* Revisión y depuración de errores 
* Edición de tablas 
* Validación clínica de resultados 
* Terminar artículo
...


# Flow chart    

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}


flow_global

flow_global2


```


# Pre-post descriptive analysis 

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

export2md(taula1, caption = paste0("Table 1: Sociodemographic variables pre matching"))

export2md(taula1.2, caption = paste0("Table 1.2 Sociodemographic variables pre matching"))

export2md(taulaPS, caption = paste0("Table 1: Sociodemographic variables post matching (N=",count(dades),")"))


```

## Covariate Balance

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

library(gridExtra)
library(patchwork)

# plot_Covariable<-gridExtra::grid.arrange(plot_ISGLT2_IDPP41, plot_ISGLT2_SU,
#                        ncol = 2,nrow=1,
#                        top=textGrob("Covariate Balance"),widths=c(3.65,2.75))

plot_ISGLT2_SU <-plot_ISGLT2_SU +  theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())

combined <- plot_ISGLT2_IDPP41 + plot_ISGLT2_SU & theme(legend.position = "bottom")
Covariable_Balance<-
  combined + plot_layout(guides = "collect")+
  plot_annotation(title = "Covariate Balance",theme = theme(plot.title = element_text(hjust = 0.5))) + 
  plot_layout(nrow = 1)

Covariable_Balance

ggsave(here::here(dir_fig,"figure_covariate.tiff"),Covariable_Balance)
ggsave(here::here(dir_fig,"figure_covariate.png"),Covariable_Balance)

```


# Descriptive analysis 

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

export2md(taules[[1]], caption = "Table 1: Sociodemographic variables")

# export2md(taules[[2]], caption = "Table 2: Baseline comorbidity variables")

export2md(taules[[3]], caption = "Clinical variables related to T2DM")

formula<-formula_compare("tabla_basal",y="grup",taulavariables = conductor_variables)
descrTable(formula,dades_completes,show.p.overall = F,show.all = T,hide.no ="No",show.n = T) %>% 
  export2md(caption = "Comorbidity profile and baseline clinical characteristics by group")

formula<-formula_compare("tbasal_continu",y="grup",taulavariables = conductor_variables)
descrTable(formula,dades_completes,show.p.overall = F,show.all = T,hide.no ="No",show.n = T, method=2, 
           extra.labels = c("","")) %>% 
  export2md(caption = "Comorbidity profile and baseline clinical characteristics by group")


# Dades que falten de seguiment 
descrTable(grup~situacio+temps_seguiment,dades_completes,method = 4,show.p.overall = F,show.all = T,hide.no ="No",show.n = T,extra.labels = c("","","")) %>% 
  export2md(caption = "Seguimiento de pacientes by group")

# Dades que falten de seguiment 
descrTable(grup~temps_seguiment,dades_completes,method = 1,show.p.overall = F,show.all = T,hide.no ="No",show.n = T,extra.labels = c("","","")) %>% 
  export2md(caption = "Seguimiento de pacientes by group")


```

# Primary objective

To compare the proportion of patients achieving the reduction in HbA1c values of at least 0.5%, a weight reduction of at least 3%, after the addition of a SU, an DPP-4i or an SGLT-2i to the treatment with metformin in patients with T2DM and insufficient glycemic control in the medium-long term, up to a maximum of 24 months of follow-up. 

## Resultado 1. (Outcomes principales)

- Descriptivo (Complete case Analysis vs Imputation cases analysis (m=10))
```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}


# 11.1 DESCRIPTIU outcomes dades completes vs dades imputades
formula_temp<-formula_compare("out1_ampliat",y="grup",taulavariables = conductor_variables)

temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c(""))
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c(""))

cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% 
   export2md(caption = "Variables related to treatment efficacy")
  

# 11.1 DESCRIPTIU outcomes dades completes vs dades imputades
formula_temp<-formula_compare("outcomes_continus",y="grup",taulavariables = conductor_variables)
temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c("",""),method = 2)
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c("",""),method = 2)

cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% 
   export2md(caption = "Variables related to treatment efficacy")
  




```


## Resultado 1. Comparativo 

- Diferencia entre grupos y Odds Ratio de reducción de HBA1C o Peso en relación a un grupo de referencia (IDPP4) crudo y ajustado
- Reducción de HBA1C en relación a un grupo de referencia (IDPP4) crudo, ajustado,  casos completos y mediante imputación múltiple de missings

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

taula_estimacions_GRUP %>% kable(digits = 3, caption = "Estimaciones en relación a un grupo de referencia (IDPP4) crudo y ajustado") %>% kableExtra::kable_styling()
print("")
taula_estimacions_IDPP4 %>% kable(digits = 3,caption = "Estimaciones del grupo IDPP4 en relación al resto ") %>% kableExtra::kable_styling()
print("")
taula_estimacions_SU %>% kable(digits = 3,caption = "Estimaciones del grupo SUFUNUREA en relación al resto ") %>% kableExtra::kable_styling()
print("")
taula_estimacions_ISGLT2 %>% kable(digits = 3,caption = "Estimaciones del grupo ISGLT2 en relación al resto ") %>% kableExtra::kable_styling()

```


## Resultado 1.bis Comparativo (respecto ISGLT2)

```{r,  message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}
# Mesures d'asociació per cada outcome 


taula_estimacions_GRUP_ref_ISGLT2 %>% kable(digits = 3, caption = "Estimaciones en relación a un grupo de referencia (ISGLT2) crudo y ajustado") %>% kableExtra::kable_styling()



```


## Forest plot 

### Diferencias estandarizadas entre grupos 2 a 2  Effect Size + IC99%

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge",fig.height=10,fig.width=14}
# Extreure outcomes principals / secundaris
llista_outcomes <- extreure.variables("outcomes1",conductor_variables) %>% tibble(outcome=.)
llista_out2 <- extreure.variables("outcomes2",conductor_variables) %>% tibble(outcome=.)
llista_out1 <- extreure.variables("outcomes1",conductor_variables) %>% tibble(outcome=.) %>% anti_join(llista_out2)


# 1. ISGLT2 vs IDPP4 (ref)
# Plotejar ISGLT2 vs IDPP4
forest_ISGLT2vsIDPP4<-forest_MET(taula_estimacions_GRUP,"grupISGLT2",titul ="DPP-4i  vs.  SGLT-2i",color=FALSE)

# 2. SU vs IDPP4 (ref)
forest_SUvsIDPP4<-forest_MET(taula_estimacions_GRUP,"grupSU",titul = "DPP-4i  vs.  SU", color=FALSE)

# 3. SU vs ISGLT2 (ref)
forest_SUvsISGLT2<-forest_MET(taula_estimacions_GRUP_ref_ISGLT2,"grup2SU",titul="SGLT-2i  vs.  SU", color=FALSE)

forest_SUvsIDPP4
forest_SUvsISGLT2
forest_ISGLT2vsIDPP4

# Fer el mateix separant primaris de secundaris
# Primaries
# Plotejar ISGLT2 vs IDPP4
forest_ISGLT2vsIDPP4_1<-forest_MET(taula_estimacions_GRUP,"grupISGLT2",titul ="DPP-4i  vs.  SGLT-2i",color=FALSE,llista_out1)
# 2. SU vs IDPP4 (ref)
forest_SUvsIDPP4_1<-forest_MET(taula_estimacions_GRUP,"grupSU",titul = "DPP-4i  vs.  SU", color=FALSE,llista_out1)
# 3. SU vs ISGLT2 (ref)
forest_SUvsISGLT2_1<-forest_MET(taula_estimacions_GRUP_ref_ISGLT2,"grup2SU",titul="SGLT-2i  vs.  SU", color=FALSE,llista_out1)

# Secundaries 
# Plotejar ISGLT2 vs IDPP4
forest_ISGLT2vsIDPP4_2<-forest_MET(taula_estimacions_GRUP,"grupISGLT2",titul ="DPP-4i vs. SGLT-2i",color=FALSE,llista_out2)
# 2. SU vs IDPP4 (ref)
forest_SUvsIDPP4_2<-forest_MET(taula_estimacions_GRUP,"grupSU",titul = "DPP-4i vs. SU", color=FALSE,llista_out2)
# 3. SU vs ISGLT2 (ref)
forest_SUvsISGLT2_2<-forest_MET(taula_estimacions_GRUP_ref_ISGLT2,"grup2SU",titul="SGLT-2i vs. SU", color=FALSE,llista_out2)


```


```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge",fig.height=10,fig.width=14}
# Combinar-ho amb un Panel plot
library(gridExtra)
library(patchwork)

forest_SUvsIDPP4 <-forest_SUvsIDPP4 +  theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
forest_SUvsISGLT2 <-forest_SUvsISGLT2 +  theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())

# forest_ISGLT2vsIDPP4<-forest_ISGLT2vsIDPP4 +  theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())

# Food note de Variables d'ajust
vector_ajust<-extreure.variables("v.ajust",conductor_variables) %>% etiquetar_vector(taulavariables = conductor_variables)
foodnote_vajust<-paste0(vector_ajust,collapse = ", ")

# Colocar food note 
forest_ISGLT2vsIDPP4<-
  forest_ISGLT2vsIDPP4+
  labs(caption=paste0("* Magnitude of effect size by Cohen's f: Small:<|0.10|; Medium:|0.10-0.25|; Large: >|0.4| \n Adjusted by ",foodnote_vajust))+
   theme(plot.caption = element_text(hjust = 0))

combined <- forest_ISGLT2vsIDPP4 + forest_SUvsIDPP4 + forest_SUvsISGLT2 & theme(legend.position = "bottom")

# forest_plot_Panel<-
#   combined + plot_layout(guides = "collect")+
#   plot_annotation(title = "Forest plot of Standardised differences between groups by outcome",theme = theme(plot.title = element_text(hjust = 0.5))) + plot_layout(nrow = 1) 

# Elimino titul
forest_plot_Panel<-combined + plot_layout(guides = "collect")

ggsave(here::here(dir_fig,"figure_ForestPlot.tiff"),forest_plot_Panel,width = 14,height = 10,dpi = 150)
ggsave(here::here(dir_fig,"figure_ForestPlot.png"),forest_plot_Panel,width = 14,height = 10)

forest_plot_Panel


```


## Forest plot només estimacions ajustades


```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge",fig.height=10,fig.width=14}

# Filtro només estimacions crues
taula_estimacions_GRUP<-taula_estimacions_GRUP %>% filter(type=="Adjusted")
taula_estimacions_GRUP_ref_ISGLT2<-taula_estimacions_GRUP_ref_ISGLT2 %>% filter(type=="Adjusted")

# 1. ISGLT2 vs IDPP4 (ref)
# Plotejar ISGLT2 vs IDPP4
forest_ISGLT2vsIDPP4<-forest_MET(taula_estimacions_GRUP,"grupISGLT2",titul ="DPP-4i  vs.  SGLT-2i",color=FALSE)

# 2. SU vs IDPP4 (ref)
forest_SUvsIDPP4<-forest_MET(taula_estimacions_GRUP,"grupSU",titul = "DPP-4i  vs.  SU", color=FALSE)

# 3. SU vs ISGLT2 (ref)
forest_SUvsISGLT2<-forest_MET(taula_estimacions_GRUP_ref_ISGLT2,"grup2SU",titul="SGLT-2i  vs.  SU", color=FALSE)

# Eliminar eix y
forest_SUvsIDPP4 <-forest_SUvsIDPP4 +  theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
forest_SUvsISGLT2 <-forest_SUvsISGLT2 +  theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())

# Food note de Variables d'ajust
vector_ajust<-extreure.variables("v.ajust",conductor_variables) %>% etiquetar_vector(taulavariables = conductor_variables)
foodnote_vajust<-paste0(vector_ajust,collapse = ", ")

# Colocar food note 
forest_ISGLT2vsIDPP4<-
  forest_ISGLT2vsIDPP4+
  labs(caption=paste0("* Magnitude of effect size by Cohen's f: Small:<|0.10|; Medium:|0.10-0.25|; Large: >|0.4| \n Adjusted by ",foodnote_vajust))+
   theme(plot.caption = element_text(hjust = 0))

combined <- forest_ISGLT2vsIDPP4 + forest_SUvsIDPP4 + forest_SUvsISGLT2 & theme(legend.position = "bottom")

forest_plot_Panel<-  combined  +  plot_layout(guides = "collect") 

forest_plot_Panel


ggsave(here::here(dir_fig,"figure_ForestPlot_Adj.tiff"),forest_plot_Panel,width = 14,height = 10,dpi = 150)
ggsave(here::here(dir_fig,"figure_ForestPlot_Adj.png"),forest_plot_Panel,width = 14,height = 10)


```



## Forest separant primaries de secundaries

```{r,message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge",fig.height=10,fig.width=14}

# Combinar-ho amb un Panel plot

forest_SUvsIDPP4_1

forest_SUvsISGLT2_1

forest_ISGLT2vsIDPP4_1


forest_SUvsIDPP4_2

forest_SUvsISGLT2_2

forest_ISGLT2vsIDPP4_2


```


# Secondary objectives

- To estimate the mean reduction of HbA1c and body weight separately after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control (up to 24 months).

## Resultado 1.2. Descriptivo  

> Reducción media 

```{r outcomes_secundaris, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

# Outcomes continus 

formula<-formula_compare(x="outcomes_continus",y="grup",taulavariables = conductor_variables)

descrTable(formula,data=dades_completes,show.p.overall = F, extra.labels=c(""), show.n = T) %>% 
  export2md(caption = "Mean reduction of HbA1c and body weight, complete cases")

descrTable(formula,data=dades %>% filter(.imp>0),show.p.overall = F,extra.labels=c(""),show.n = T) %>% 
  export2md(caption = "Mean reduction of HbA1c and body weight, imputation cases")

formula<-formula_compare(x="outcomes_continus",y="grup",taulavariables = conductor_variables)

descrTable(formula,data=dades_completes,show.p.overall = F, extra.labels=c("",""), show.n = T,method = 2) %>% 
  export2md(caption = "Mean reduction of HbA1c and body weight, complete cases")

descrTable(formula,data=dades %>% filter(.imp>0),show.p.overall = F,extra.labels=c("",""),show.n = T,method = 2) %>% 
  export2md(caption = "Mean reduction of HbA1c and body weight, imputation cases")


```

-----

## Resultado 2. (Outcomes secundarios)

- To assess the average reduction in systolic blood pressure after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control.

- Descriptivo (Complete case Analysis vs Imputation cases analysis (m=10))

### Resultado 2. Descriptivo 

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}
# Ara fer-ho de cada grup respecte la resta 

# 1. DESCRIPTIVA (Media)
formula_temp<-formula_compare("table6",y="grup",taulavariables = conductor_variables)
temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c(""))
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c(""))
cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% export2md()


# 1. DESCRIPTIVA (Mediana)
formula_temp<-formula_compare("table6",y="grup",taulavariables = conductor_variables)
temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c("","",""),method=2)
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c("","",""),method=2)
cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% export2md()


```


### Resultado 2. Comparativo 

- Diferencia entre grupos en relación a un grupo de referencia (IDPP4) crudo y ajustado
- Reducción de HBA1C en relación a un grupo de referencia (IDPP4) crudo, ajustado,  casos completos y mediante imputación múltiple de missings

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

taula_estimacions2_GRUP %>% kable(digits = 3, caption = "Estimaciones en relación a un grupo de referencia (IDPP4) crudo y ajustado") %>% kableExtra::kable_styling()
print("")
taula_estimacions2_IDPP4 %>% kable(digits = 3,caption = "Estimaciones del grupo IDPP4 en relación al resto ") %>% kableExtra::kable_styling()
print("")
taula_estimacions2_SU %>% kable(digits = 3,caption = "Estimaciones del grupo SUFUNUREA en relación al resto ") %>% kableExtra::kable_styling()
print("")
taula_estimacions2_ISGLT2 %>% kable(digits = 3,caption = "Estimaciones del grupo ISGLT2 en relación al resto ") %>% kableExtra::kable_styling()

```


## Resultado 3. (Outcomes secundarios)

- To assess the evolution of lipid profile (mean total cholesterol, HDL, LDL, and triglycerides) after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient control glycemic. 

### Resultado 3.1. Descriptivo 

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

formula_temp<-formula_compare("table7",y="grup",taulavariables = conductor_variables)
temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c(""))
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c(""))
cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% export2md()



formula_temp<-formula_compare("table7",y="grup",taulavariables = conductor_variables)
temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c("",""),method = 2)
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c("",""),method = 2)
cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% export2md()


```

### Resultado 3.2. Comparativo 
```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

taula_estimacions3_GRUP %>% kable(digits = 3, caption = "Estimaciones en relación a un grupo de referencia (IDPP4) crudo y ajustado") %>% kableExtra::kable_styling()
print("")
taula_estimacions3_IDPP4 %>% kable(digits = 3,caption = "Estimaciones del grupo IDPP4 en relación al resto ") %>% kableExtra::kable_styling()
print("")
taula_estimacions3_SU %>% kable(digits = 3,caption = "Estimaciones del grupo SUFUNUREA en relación al resto ") %>% kableExtra::kable_styling()
print("")
taula_estimacions3_ISGLT2 %>% kable(digits = 3,caption = "Estimaciones del grupo ISGLT2 en relación al resto ") %>% kableExtra::kable_styling()

```


## Resultado 4 (Seguimiento de fármacos dispensados: MPR, adherencia y stops).

### Resultado 4.1. Descriptiva

- To assess  the adherence to treatment after the addition of an SU, a DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control.


```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

formula_temp<-formula_compare("table8",y="grup",taulavariables = conductor_variables)
temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c(""),show.all = T)
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c(""),show.all = T)
cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% export2md()

formula_temp<-formula_compare("table8",y="grup",taulavariables = conductor_variables)
temp_comp<-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp==0),show.p.overall = F,show.n = T,extra.labels =c("",""),method = 2,show.all = T)
# Descriptivo datos imputados:
temp_imp <-compareGroups::descrTable(formula_temp,data=dades %>% filter(.imp>0),show.p.overall = F,show.n = T,extra.labels =c("",""),method = 2,show.all = T)
cbind("Complete case Analysis" = temp_comp, "Imputation cases analysis (m=10)" = temp_imp) %>% export2md()



```

### Resultado 4.1. Comparativa

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

# Afegeixo etiqueta a outcome 
taula_adhrecia_GRUP<-etiquetar_taula(taula_adhrecia_GRUP,camp="outcome",conductor_variables)

# COMPARATIVA
taula_adhrecia_GRUP %>% kable(digits = 3,caption = "Estimaciones del de cada grupo en relqación a IDPP4 y resto, crudas y ajustadas") %>% kableExtra::kable_styling()

```

### Resultado 4.2. 

- To assess  the percentage of suspensions/dropouts of treatment at 6, 12 and 24 months.

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}


cfit<-survfit(Surv(temps.STOP.FD,STOP.FD)~grup,dades_completes)

taula_percentatge_stops<-broom::tidy(cfit) %>% filter(time==180) %>% 
  bind_rows(broom::tidy(cfit) %>% filter(time==365)) %>% 
  bind_rows(broom::tidy(cfit) %>% filter(time==730)) %>% 
  transmute(strata,time_in_days=time,estimate_stop=1-estimate,CI_linf=1-conf.high,CI_sup=1-conf.low) %>% 
  kable(caption = "Percentage of suspensions/dropouts of treatment at 6, 12 and 24",digits = 3) %>% kableExtra::kable_styling()

taula_percentatge_stops

survminer::ggsurvplot(cfit, data = dades_completes,
                      main = "Survival curve",
                      title= "Suspensions/dropouts during follow-up before 24 months",
                      legend.labs = c("IDPP4","SGLT-2i","SU"),
                      size = 0.5,
                      ylim = c(0,1),
                      xlim = c(0,730),
                      break.x.by=180,
                      xlab = "Time in days",
                      risk.table = F,
                      censor.shape="|", censor.size = 1)


survminer::ggsurvplot(cfit, data = dades_completes,
                      legend.labs = c("IDPP4","SGLT-2i","SU"),
                      main = "Survival curve",
                      title= "Suspensions/dropouts during follow-up before 24 months",
                      size = 0.5,
                      ylim = c(0,1),
                      xlim = c(0,730),
                      break.x.by=180,
                      xlab = "Time in days",
                      risk.table = T,
                      censor.shape="|", censor.size = 1)

descrTable(temps_fins_stop~grup,data=dades_completes,show.ratio = T) %>% 
  export2md()


```


### Resultado 5. 

- To describe the relevant adverse reactions produced after the addition of an SU, an DPP-4i or an SGLT-2i in patients with DM2 treated with metformin and insufficient glycemic control.


```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

# Generar taula
formula_temp<-formula_compare("table10",y="grup",taulavariables = conductor_variables)

compareGroups::descrTable(formula_temp,data=dades_completes,show.p.overall = F,hide = T,show.all = T) %>% export2md()




```


### Resultado 6. 

- To describe the comorbidity profile and baseline clinical characteristics of patients with DM2 and insufficient control with metformin in monotherapy after addition of a treatment with SU, DPP-4i or SGLT-2i.

- Se ha considerado grupo con insuficiente control después de la addición, una HBA1C >=7%  después del tratamiento antes de los 24 meses. 

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

dades_temp<-dades_completes %>% filter(HBA1C.valor324m>=7)
formula<-formula_compare("tabla_basal",y="grup",taulavariables = conductor_variables)
descrTable(formula,dades_temp,show.p.overall = T,show.all = T) %>% 
  export2md(caption = "Comorbidity profile and baseline clinical characteristics by group")



```

- To describe the clinical characteristics of patients who have the best efficacy results (patients with a reduction of HbA1c of more than 0.5% and a weight reduction of at least 3% before the end of follow-up).

- Comparativa de características basales de pacientes que consiguen una reducción reduction de peso y HBA1C del 3% y de 0.5 unidades de HBA1c(%) en algun momento antes de los 24meses

### Resultado 7. 

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}


formula<-formula_compare("tabla_basal",y="PESHB.dif324m.cat",taulavariables = conductor_variables)
descrTable(formula,dades_completes,show.p.overall = T) %>% 
  export2md(caption = "Clinical characteristics of patients who have the best results (Weight-HBA1C reduction vs No) before the end of follow-up")



```

### Resultado 8.

Frequencia de cada fármaco por grupo

```{r, message=FALSE, warning=FALSE, include=T, echo=FALSE,size="huge"}

# Capturar grupindex i filtrar 
dt_codindex<-dt_codindex %>% mutate(grup=if_else(grup_index=="iSGLT2","ISGLT2",grup_index))

dt_temp<-dades_completes %>% transmute(idp=as.character(idp),grup=as.character(grup2)) %>% 
  left_join(dt_codindex, by=c("idp","grup"))

# Taula 
table1::table1(~cod_index+fx_index | grup,data=dt_temp)


```
